// src/interfaces/response/movies-response.ts
export interface MoviesResponse {
  data: Movie[];
  totalAmount: number;
  offset: number;
  limit: number;
}

export interface Movie {
  title: string;
  tagline: string;
  vote_average: number;
  vote_count: number;
  release_date: string;
  poster_path: string;
  overview: string;
  budget: number;
  revenue: number;
  runtime: number;
  genres: string[];
  id: number;
}